feat: move shared code to util module and require identityRef
